change layer int width

+++ app/src/world_edit.hpp
@@ -294,7 +294,7 @@ private:
       world::planning_connection_id last_planning_connection = world::max_id;
       world::boundary_id last_boundary = world::max_id;
 
-      int last_layer = 0;
+      int16 last_layer = 0;
 
       container::ring_set<lowercase_string, 8> last_used_object_classes;
    } _last_created_entities;
@@ -390,7 +390,7 @@ private:
    world::path_id _move_entire_path_id = {};
    world::sector_id _move_sector_point_id = {};
    std::size_t _move_sector_point_index = 0;
-   int _selection_set_layer = 0;
+   int16 _selection_set_layer = 0;
 
    float2 _select_start_position;
