+++ src/world/hintnode.hpp
@@ -50,7 +50,7 @@ enum class hintnode_mode {
 
 struct hintnode {
    std::string name;
-   int layer = 0;
+   int16 layer = 0;
 
    quaternion rotation = {1.0f, 0.0f, 0.0f, 0.0f};
    float3 position{};
